requirements.txt
pandas==2.2.2
psycopg2==2.9.9
python-dotenv==1.0.1
